Minor bug fixes.
Note:
----------------------------------------------------------------------------
cannot have leading zeroes in zip because if the code:
  def zip
      "%05d" % read_attribute(:zip)
  end
is in the model, then I get an error 'can't covert nill to integer' if I enter an empty zip. Even if I validate :zip with :presence => true.
There is no solution to this, the way we add leading zeroes is terrible. Need a better solution to this.
----------------------------------------------------------------------------